add jwt logic

+++ src/main/java/com/smartspend/controller/UserController.java
@@ -4,6 +4,9 @@ package com.smartspend.controller;
 import com.smartspend.model.User;
 import com.smartspend.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,23 +20,40 @@ public class UserController {
     private UserService userService;
 
     @PostMapping
-    public User createUser(@RequestBody User user) {
-        return userService.createUser(user);
+//    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<String> createUser(@RequestBody User user) {
+        if(userService.getUserByEmail(user.getEmail()) != null) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("User already exists");
+        }
+
+      User user1 = userService.createUser(user);
+        return ResponseEntity.status(HttpStatus.CREATED).body(user1.toString());
     }
 
     @GetMapping
+//    @PreAuthorize("hasRole('ADMIN')")
     public List<User> getAllUsers() {
         return userService.getAllUsers();
     }
 
     @GetMapping("/{id}")
+//    @PreAuthorize("hasRole('ADMIN')")
     public User getUser(@PathVariable int id) {
         return userService.getUserById(id);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id) {
+//    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<String> deleteUser(@PathVariable int id) {
+
+        if(userService.getUserById(id) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User with id " + id + " does not exist");
+
+
+        }
         userService.deleteUserById(id);
+        return ResponseEntity.ok().body("User deleted successfully");
 }
 
 }
